requirements.txt
streamlit
tensorflow
numpy
pillow
opencv-python-headless
streamlit-folium
folium
